pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-irctc_booking"
version = "0.1.0"
description = "IRCTC booking simulation environment for OpenEnv"
requires-python = ">=3.11"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.3",
    "openai>=1.10.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.0.0",
    "hypothesis>=6.98.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m irctc_booking.server.app
server = "irctc_booking.server.app:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["irctc_booking*"]